impl hand estimation model and its converter
