+++ README.md
@@ -41,10 +41,12 @@ Execute the following commands to download the trained model and convert it to n
 
 ```
 cd models
-wget http://posefs1.perception.cs.cmu.edu/Users/ZheCao/pose_iter_440000.caffemodel
+wget http://posefs1.perception.cs.cmu.edu/OpenPose/models/pose/coco/pose_iter_440000.caffemodel
 wget http://posefs1.perception.cs.cmu.edu/OpenPose/models/face/pose_iter_116000.caffemodel
-python convert_pose_model.py pose_iter_440000.caffemodel
-python convert_face_model.py pose_iter_116000.caffemodel
+wget http://posefs1.perception.cs.cmu.edu/OpenPose/models/hand/pose_iter_102000.caffemodel
+python convert_model.py posenet pose_iter_440000.caffemodel coco_posenet.npz
+python convert_model.py facenet pose_iter_116000.caffemodel facenet.npz
+python convert_model.py handnet pose_iter_102000.caffemodel handnet.npz
 cd ..
 ```
